requirements.txt
streamlit
pandas
sqlalchemy
requests
beautifulsoup4
extruct
w3lib
trafilatura
readability-lxml
lxml
pytesseract
pillow
pdfplumber
python-slugify
python-dateutil